added a simple method to provide for issue #11, build with older containers, reqs any FORCE_BUILD to provide <vars>

+++ get-latest-container-ver-num.sh
@@ -7,8 +7,11 @@ index=0
 for pkg in ${pkgs[@]}; do
 	#echo ${vars[$index]}
 
-		
-	eval "${vars[$index]}=`curl https://api.bintray.com/packages/esaude/$pkg/versions/_latest | grep -P '(?<=\"name\":\").*?(?=",)' -o`"
-		
+	if [ -z "$FORCE_BUILD" ] || [ "$FORCE_BUILD" == 0 ]; then	
+		eval "${vars[$index]}=`curl https://api.bintray.com/packages/esaude/$pkg/versions/_latest | grep -P '(?<=\"name\":\").*?(?=",)' -o`"
+	else
+		eval "tmp_var=\$${vars[$index]}"
+		echo "using exisiting environment variable \$${vars[$index]}=$tmp_var"
+	fi
 	let "index++"
 done
